Réorganisation structurelle du projet -> blocs visuels
