+++ src/list_interface.py
@@ -4,8 +4,11 @@ Ce fichier contient l'ensemble des ressources liées uniquement
 """
 
 import pygame
-import pygame_textinput
 import general
+
+from control_bar import ControlBar
+from grid import Grid
+from task_editor import TaskEditor
 
 
 class TaskList(general.Div):
@@ -131,319 +134,4 @@ class TaskList(general.Div):
         self.grid.display(self.surface)
         if self.editor is not None:
             self.editor.display(self.surface)
-        screen.blit(self.surface, self.hit_box)
-
-# ============================================================================
-
-
-class ControlBar(general.Div):
-    """
-    Barre de contrôles, elle permet de réaliser
-    différentes action à l'aide de boutons
-    """
-    def __init__(self, size: tuple, pos: tuple):
-        super().__init__(size, pos)
-        self.surface.fill((0, 0, 255))
-
-        # Texte - Exploitable par pygame
-        self.text = general.get_screen_text_for("Control Bar", size[1] // 5)
-        self.text_rect = self.text.get_rect()
-        self.text_rect.center = (
-            self.surface.get_width() // 2,
-            self.surface.get_height() // 2
-        )
-
-        # Bouton - 'Ajouter'
-        size_b = (size[0] // 10, 2 * size[1] // 3)
-        self.ajouter = general.Button(size_b, (0, 0), "ajouter")
-        self.ajouter.hit_box.right = 19 * size[0] // 20
-        self.ajouter.hit_box.centery = size[1] // 2
-
-    def update(self, pos: tuple) -> str:  # -------------------------
-        if self.ajouter.is_under(pos):
-            return "ajouter"
-        return ""
-
-    def display(self, screen: pygame.Surface):  # -------------------
-        self.surface.blit(self.text, self.text_rect)
-        self.ajouter.display(self.surface)
-        screen.blit(self.surface, self.hit_box)
-
-# ============================================================================
-
-
-class Grid(general.Div):
-    """
-    Grille d'affichage des tâches, elle permet d'afficher et de manager
-    les tâches auquelles elle est associée en leur donnant une taille identique
-    et une position dynamique.
-    """
-    def __init__(self, size: tuple, pos: tuple, layout_size: tuple):
-        super().__init__(size, pos)
-        self.elements = []
-        self.gap = size[0] // 100
-
-        # Dimensions de la grille
-        self.nb_colonne = layout_size[0]
-        self.nb_ligne = layout_size[1]
-        self.tile_size = (
-            (size[0] - self.gap) // self.nb_colonne - self.gap,
-            (size[1] - self.gap) // self.nb_ligne - self.gap
-        )
-
-    def fill_elements(self, new_list: list):  # ---------------------
-        """ Construction dynamique des tâches """
-        self.elements = []
-        i = 0
-        for elt in new_list:
-            pos = (
-                self.gap + (i % self.nb_colonne) * (self.tile_size[0] + self.gap),
-                self.gap + (i // self.nb_colonne) * (self.tile_size[1] + self.gap)
-            )
-            task = Task(self.tile_size, pos, elt)
-            self.elements.append(task)
-            i += 1
-
-    def display(self, screen: pygame.Surface):  # -------------------
-        self.surface.fill((0, 255, 0))
-        for elt in self.elements:
-            elt.display(self.surface)
-        screen.blit(self.surface, self.hit_box)
-
-# ============================================================================
-
-
-class Task(general.Div):
-    def __init__(self, size: tuple, pos: tuple, spec: dict):
-        super().__init__(size, pos)
-        self.spec = spec
-        self.surface.fill((0, 0, 50))
-
-        # Éléments
-        self.ended = general.CheckBox(
-            (size[1] // 5, size[1] // 5),
-            (0, 0),
-            spec["fini"]
-        )
-        self.delete = general.Button(
-            (size[0] // 3, size[1] // 5),
-            (size[1] // 5 + 5, 0),
-            "supprimer"
-        )
-        self.nom = general.Text(
-            (9 * size[0] // 10, 2 * size[1] // 3),
-            (size[0] // 20, size[1] // 5 + 5),
-            spec["nom"],
-            size[1] // 7
-        )
-        self.effort = general.get_screen_text_for(
-            str((spec["difficulté"] + spec["longueur"])/2), size[1] // 5
-        )
-        self.effort_rect = self.effort.get_rect()
-        self.effort_rect.topright = (size[0], 0)
-
-    def update(self, pos: tuple) -> str:  # -------------------------
-        t_pos = self.get_relative_pos(pos)
-
-        # Checkbox
-        if self.ended.is_under(t_pos):
-            self.ended.switch_state()
-            self.spec["fini"] = self.ended.get_state()
-
-        # Delete
-        elif self.delete.is_under(t_pos):
-            return "delete"
-
-        elif self.nom.is_under(t_pos):
-            return "editor"
-
-        return ""
-
-    def display(self, screen: pygame.Surface):  # -------------------
-        self.ended.display(self.surface)
-        self.delete.display(self.surface)
-        self.nom.display(self.surface)
-        self.surface.blit(self.effort, self.effort_rect)
-        screen.blit(self.surface, self.hit_box)
-
-# ============================================================================
-
-
-class TaskEditor(general.Div):
-    def __init__(self, size: tuple, pos: tuple, task: dict):
-        super().__init__(size, pos)
-        self.task = task
-        self.surface.fill((255, 0, 0))
-        self.selected_area = None
-
-        # Enregistrer
-        self.save = general.Button(
-            (100, 50),
-            (0, 0),
-            "enregistrer"
-        )
-
-        # Quitter
-        self.quit = general.Button(
-            (100, 50),
-            (100, 0),
-            "quitter"
-        )
-
-        # Nom
-        self.nom_input = pygame_textinput.TextInputManager()
-        self.nom_visualizer = general.InputVisualizer(
-            (size[0], size[1] // 8),
-            (0, 100),
-            3
-        )
-
-        # Description
-        self.description_input = pygame_textinput.TextInputManager()
-        self.description_visualizer = general.InputVisualizer(
-            (size[0], size[1] // 4),
-            (0, self.nom_visualizer.hit_box.bottom + 10),
-            7
-        )
-
-        # Difficulté
-        self.difficulty = general.Scale(
-            (size[0], size[1] // 10),
-            (0, self.description_visualizer.hit_box.bottom + 30),
-            5
-        )
-
-        # Longueur
-        self.longueur = general.Scale(
-            (size[0], size[1] // 10),
-            (0, self.difficulty.hit_box.bottom + 20),
-            5
-        )
-
-        # Checkbox
-        self.checkbox = general.CheckBox(
-            (size[1] // 10, size[1] // 10),
-            (0, self.longueur.hit_box.bottom + 30),
-        )
-
-    def set_task(self):
-        self.task["nom"] = self.nom_visualizer.raw_text
-        self.task["description"] = self.description_visualizer.raw_text
-        self.task["difficulté"] = self.difficulty.value
-        self.task["longueur"] = self.longueur.value
-        self.task["fini"] = self.checkbox.state
-
-    def set_info(self):
-
-        # Nom
-        self.nom_input.left = self.task["nom"]
-        self.nom_visualizer.change_text(self.nom_input.left, "")
-        self.nom_visualizer.screen_text.pop()
-
-        # Description
-        self.description_input.left = self.task["description"]
-        self.description_visualizer.change_text(
-            self.description_input.left, ""
-        )
-        self.description_visualizer.screen_text.pop()
-
-        # Difficulté
-        self.difficulty.value = self.task["difficulté"]
-        self.difficulty.set_elements()
-
-        # Longueur
-        self.longueur.value = self.task["longueur"]
-        self.longueur.set_elements()
-
-        # Checkbox
-        self.checkbox.state = self.task["fini"]
-        self.checkbox.color_state()
-
-    def update_text(self, events: pygame.event.Event):
-        if self.selected_area is not None:
-            if self.selected_area == "nom":
-                self.nom_input.update(events)
-                self.nom_visualizer.change_text(
-                    self.nom_input.left,
-                    self.nom_input.right
-                )
-            elif self.selected_area == "description":
-                self.description_input.update(events)
-                self.description_visualizer.change_text(
-                    self.description_input.left,
-                    self.description_input.right
-                )
-
-    def unselect_nom(self):
-        if (
-            self.nom_visualizer.visible
-            and self.nom_visualizer.screen_text
-        ):
-            self.nom_visualizer.screen_text.pop()
-        self.nom_visualizer.screen_text = self.nom_visualizer.wrap(
-            self.nom_visualizer.raw_text
-        )
-
-    def unselect_description(self):
-        if (
-            self.description_visualizer.visible
-            and self.description_visualizer.screen_text
-        ):
-            self.description_visualizer.screen_text.pop()
-        self.description_visualizer.screen_text = self.description_visualizer.wrap(
-            self.description_visualizer.raw_text
-        )
-
-    def update(self, pos: tuple) -> str:
-        if self.save.is_under(pos):
-            return "enregistrer"
-        elif self.quit.is_under(pos):
-            return "quitter"
-        elif self.difficulty.is_under(pos):
-            rel_pos = self.difficulty.get_relative_pos(pos)
-            self.difficulty.update(rel_pos)
-        elif self.longueur.is_under(pos):
-            rel_pos = self.longueur.get_relative_pos(pos)
-            self.longueur.update(rel_pos)
-        elif self.checkbox.is_under(pos):
-            self.checkbox.switch_state()
-        else:
-
-            # Sélection Nom
-            if self.nom_visualizer.is_under(pos):
-                self.selected_area = "nom"
-                self.nom_input.left = self.nom_visualizer.raw_text
-                self.nom_input.right = ""
-                self.unselect_description()
-
-            # Sélection Description
-            elif self.description_visualizer.is_under(pos):
-                self.selected_area = "description"
-                self.description_input.left = self.description_visualizer.raw_text
-                self.description_input.right = ""
-                self.unselect_nom()
-
-            # Déselection
-            else:
-
-                # Nom
-                if self.selected_area == "nom":
-                    self.unselect_nom()
-
-                # Description
-                elif self.selected_area == "description":
-                    self.unselect_description()
-
-                self.selected_area = None
-
-        return ""
-
-    def display(self, screen: pygame.surface):
-        self.save.display(self.surface)
-        self.quit.display(self.surface)
-        self.nom_visualizer.display(self.surface)
-        self.description_visualizer.display(self.surface)
-        self.difficulty.display(self.surface)
-        self.longueur.display(self.surface)
-        self.checkbox.display(self.surface)
         screen.blit(self.surface, self.hit_box)
